requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
h11==0.12.0
idna==3.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.3
requests==2.26.0
six==1.16.0
tornado==6.1
tzlocal==3.0
urllib3==1.26.7
uvicorn==0.15.0
